+++ CHANGELOG.md
@@ -6,7 +6,22 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- Nothing.
+- [#55](https://github.com/zendframework/zend-expressive-swoole/pull/55) adds a new configuration key, `zend-expressive-swoole.swoole-http-server.logger.logger-name`.
+  It allows a custom service name which resolves to a `Psr\Log\LoggerInterface`
+  instance to be provided, in order to be wrapped in the
+  `Zend\Expressive\Swoole\Log\Psr3AccessLogDecorator`:
+
+  ```php
+  return [
+      'zend-expressive-swoole' => [
+          'swoole-http-server' => [
+              'logger' => [
+                  'logger-name' => 'my_logger',
+              ],
+          ],
+      ],
+  ];
+  ```
 
 ### Changed